% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{Bisection Algorithm}
\usage{
bisection(x_lo, x_hi, f, mid, dist, tol)
}
\arguments{
\item{x_lo}{A finite lower bound for \eqn{x} which should occur before the step.}

\item{x_hi}{A finite upper bound for \eqn{x} which should occur after the step.}

\item{f}{The function \eqn{f(x)}.}

\item{mid}{A midpoint function which operates on two points \code{(x,y)}
where \code{x <= y}.}

\item{dist}{A distance function which operates on two points \code{(x,y)}
where \code{x <= y}.}

\item{tol}{Stop when \code{dist(x_lo, x_hi) < tol}.}
}
\description{
Given an increasing step function \eqn{f(x)}, locate the point where the
step occurs.
}
