% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DGeom}
\alias{DGeom}
\alias{d_dgeom}
\alias{p_dgeom}
\alias{r_dgeom}
\alias{q_dgeom}
\title{Double Geometric Distribution}
\usage{
d_dgeom(x, p, take_log)

p_dgeom(x, p)

r_dgeom(n, p)

q_dgeom(q, p)
}
\arguments{
\item{x}{A vector of points to evaluate.}

\item{p}{probability parameter.}

\item{take_log}{If \code{TRUE} return the log-density.}

\item{n}{Number of draws to generate.}

\item{q}{A vector of probabilities to evaluate.}
}
\description{
Functions for the Double Geometric distribution.
}
\details{
Here we assume Double Geometric distribution with density
\deqn{
f(x) = \frac{\rho}{2 - \rho} (1 - \rho)^{|x|}, \quad x \in \mathbb{Z}.
}
}
