% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct-sampler-ar.R
\name{direct-sampler-ar}
\alias{direct-sampler-ar}
\alias{direct_sampler_ar}
\title{Direct sampler with accept-reject}
\usage{
direct_sampler_ar(
  n,
  w,
  g,
  tol = 1e-08,
  N = 10,
  max_rejections,
  fill_method = "small_rects",
  priority_weight = 1/2,
  verbose = FALSE
)
}
\arguments{
\item{n}{Number of draws to generate.}

\item{w}{A weight function object.}

\item{g}{A base distribution object.}

\item{tol}{Tolerance for step function approximation in customized sampler.}

\item{N}{Number of knots to use in approximation for \eqn{p(u)}. Specifically,
N+1 points will be used.}

\item{max_rejections}{Maximum number of rejections to allow. If this number
is reached, an exception will be thrown.}

\item{fill_method}{Knot selection method for customized sampler. See
\code{\link{Stepdown}}.}

\item{priority_weight}{TBD}

\item{verbose}{Return a list with the sample \code{x} and the number of
rejections \code{rejections}.}
}
\value{
A vector of length \code{n} which represents the sample.
}
\description{
Direct sampler with accept-reject
}
\details{
Direct sampling with accept-reject algorithm can be used to take exact draws
from the target distribution, with the added cost that same candidate draws
will be rejected.

The function \code{direct_sampler_ar} does not adapt the underlying step
function using rejections; here, the number of knots \code{N} is kept
constant throughout sampling.
}
