% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct-sampler.R
\name{direct-sampler}
\alias{direct-sampler}
\alias{direct_sampler_basic}
\alias{direct_sampler}
\title{Direct Sampler}
\usage{
direct_sampler_basic(n, w, g, N = 100, max_iter = 10000)

direct_sampler(
  n,
  w,
  g,
  tol = 1e-08,
  N = 100,
  fill_method = "small_rects",
  priority_weight = 1/2
)
}
\arguments{
\item{n}{Number of draws to generate.}

\item{w}{A weight function object.}

\item{g}{A base distribution object.}

\item{N}{Number of knots to use in approximation for \eqn{p(u)}. Specifically,
N+1 points will be used.}

\item{max_iter}{Maximum number of accept/reject attempts to try for each draw.}

\item{tol}{Tolerance for step function approximation in customized sampler.}

\item{fill_method}{Knot selection method for customized sampler. See
\code{\link{Stepdown}}.}

\item{priority_weight}{TBD}
}
\description{
Basic and customized versions of the direct sampler.
}
