% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct-sampler.R
\name{DirectSampler}
\alias{DirectSampler}
\alias{direct_sampler_basic}
\alias{direct_sampler}
\title{Direct Sampler}
\usage{
direct_sampler_basic(n, w, g, N = 100, max_iter = 10000)

direct_sampler(n, w, g, tol = 1e-08, N = 100, fill_method = "small_rects")
}
\arguments{
\item{n}{Number of draws to generate.}

\item{w}{An weight function object.}

\item{g}{An base distribution object.}

\item{N}{Number of knots to use in approximation for \eqn{p(u)}.}

\item{max_iter}{Maximum number of accept/reject attempts to try for each draw.}

\item{tol}{Tolerance for step function approximation in customized sampler.}

\item{fill_method}{Knot selection method for customized sampler. See
\link{Step}.}
}
\description{
Basic and customized versions of the direct sampler.
}
\details{
The object \code{w} represents a unimodal weight function. It is
expected to contain several members:
\itemize{
\item \code{log_c}: the logarithm of the value c, which is the mode of the
weight function.
\item \code{roots(log_a)}: return the roots of the equation
\eqn{\log w(x) - \rm{log_a} = 0}.
\item \code{eval(x, log = TRUE)}: evaluate the function. Return the value
on the log-scale if \code{log = TRUE}.
}

The object \code{g} represents a base distribution. It is
expected to contain two member functions:
\itemize{
\item \code{pr_interval(x1, x2)}: Return Pr(x1 < X <= x2) under
distribution g.
\item \code{r_truncated(n, x1, x2)}: Take n draws from distribution g
truncated to the interval (x1, x2].
}
}
