% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{Dirichlet}
\alias{Dirichlet}
\alias{rdirichlet}
\alias{ddirichlet}
\title{Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)

ddirichlet(x, alpha, log = FALSE)
}
\arguments{
\item{n}{Number of draws to generate}

\item{alpha}{k-dimensional vector of intensity parameters.}

\item{x}{An n-by-k matrix whose rows are points to evaluate.}

\item{log}{If \code{TRUE} return the log-density.}
}
\description{
Density and drawing function for Dirichlet distribution on the \eqn{k}
dimensional probability simplex.
}
\details{
Here we assume Dirichlet distribution with density
\deqn{
f(x) = \frac{x_1^{\alpha_1-1} \cdots x_k^{\alpha_k-1}}{B(\bm{\alpha})}.
}
}
