% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Discrete Normal}
\alias{Discrete Normal}
\alias{hi_dscnorm}
\alias{d_dscnorm}
\alias{p_dscnorm}
\alias{r_dscnorm}
\alias{q_dscnorm}
\title{Discrete Normal Distribution}
\usage{
hi_dscnorm(sigma, tol)

d_dscnorm(x, sigma, tol, take_log, normalize)

p_dscnorm(x, sigma, tol)

r_dscnorm(n, sigma)

q_dscnorm(q, sigma, tol)
}
\arguments{
\item{sigma}{Standard deviation parameter.}

\item{tol}{Tolerance to truncate to a finite support (see details).}

\item{x}{A vector of points to evaluate.}

\item{take_log}{If \code{TRUE} return the log-density.}

\item{normalize}{If \code{TRUE}, normalize the density (approximately).}

\item{n}{Number of draws to generate.}

\item{q}{A vector of probabilities to evaluate.}
}
\description{
Functions for the Discrete Normal distribution with mean zero.
}
\details{
Here we assume Discrete Normal distribution with density
\deqn{
f(x) \propto \exp( -x^2 / (2 \sigma^2) ), \quad x \in
\{ \ldots, -1, 0, 1, \ldots \}.
}

The function \code{hi_dscnorm} returns an \eqn{x} so that \eqn{[-x, x]}
contains \code{1 - tol} mass of the distribution. This is used to truncate
the distribution to have finite support for the simple implementation here.
}
\references{
Clement L. Canonne, Gautam Kamath, Thomas Steinke (2021), The Discrete
Gaussian for Differential Privacy. \url{https://arxiv.org/abs/2004.00010}.
}
