% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{Inverse Gamma}
\alias{Inverse Gamma}
\alias{rinvgamma}
\alias{dinvgamma}
\title{Inverse Gamma Distribution}
\usage{
rinvgamma(n, a, b)

dinvgamma(x, a, b, log = FALSE)
}
\arguments{
\item{n}{Number of draws to generate}

\item{a}{shape parameter.}

\item{b}{rate parameter.}

\item{x}{A vector of n points to evaluate.}

\item{log}{If \code{TRUE} return the log-density.}
}
\description{
Density and drawing function for the Inverse Gamma distribution.
}
\details{
Here we assume Inverse Gamma distribution with density
\deqn{
f(x) = \frac{b^a}{\Gamma(a)} x^{-(a+1)} e^{-b/x}.
}
}
