% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Laplace}
\alias{Laplace}
\alias{d_laplace}
\alias{p_laplace}
\alias{r_laplace}
\alias{q_laplace}
\title{Laplace Distribution}
\usage{
d_laplace(x, mu, lambda, take_log)

p_laplace(x, mu, lambda)

r_laplace(n, mu, lambda)

q_laplace(q, mu, lambda)
}
\arguments{
\item{x}{A vector of points to evaluate.}

\item{mu}{Location parameter.}

\item{lambda}{Scale parameter.}

\item{take_log}{If \code{TRUE} return the log-density.}

\item{n}{Number of draws to generate.}

\item{q}{A vector of probabilities to evaluate.}
}
\description{
Functions for the Laplace distribution.
}
\details{
Here we assume Laplace distribution with density
\deqn{
f(x) = \frac{1}{2\lambda} e^{-|x| / \lambda}.
}
}
