% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{MVN}
\alias{MVN}
\alias{rmvnorm}
\alias{rmvnorm_prec}
\alias{dmvnorm}
\alias{dmvnorm_prec}
\title{Multivariate Normal Distribution}
\usage{
rmvnorm(n, mu, Sigma)

rmvnorm_prec(n, mu, Omega)

dmvnorm(x, mu, Sigma, log = FALSE)

dmvnorm_prec(x, mu, Omega, log = FALSE)
}
\arguments{
\item{n}{Number of draws to generate.}

\item{mu}{k-dimensional vector representing the mean.}

\item{Sigma}{k-by-k covariance matrix.}

\item{Omega}{k-by-k precision matrix.}

\item{x}{An n-by-k matrix whose rows are points to evaluate.}

\item{log}{If \code{TRUE}, return the log-density.}
}
\description{
Density and drawing functions for Multivariate Normal distribution on
\eqn{\mathbb{R}^k}, parameterized either by the covariance or precision
matrix.
}
