% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{Stepdown}
\alias{Stepdown}
\title{Stepdown class}
\description{
Stepdown class

Stepdown class
}
\details{
This R6 class represents a step function approximation to the non-increasing
function \eqn{\rm{Pr}(A_u)} for \eqn{u \in [0,1]}.

The object \code{w} represents a unimodal weight function. It is
expected to contain several members:
\itemize{
\item \code{log_c}: the logarithm of the value c, which is the mode of the
weight function.
\item \code{roots(log_a)}: return the roots of the equation
\eqn{\log w(x) - \rm{log_a} = 0}.
\item \code{eval(x, log = TRUE)}: evaluate the function. Return the value
on the log-scale if \code{log = TRUE}.
}

The object \code{g} represents a base distribution. It is
expected to contain two member functions:
\itemize{
\item \code{pr_interval(x1, x2)}: Return \eqn{\rm{Pr}(x1 < X < x2)} under
distribution g.
\item \code{r_truncated(n, x1, x2)}: Take n draws from distribution g
truncated to the open interval \eqn{(x1, x2)}.
}

We make use of the \code{fibonacci_heap} structure in the
\code{datastructures} package to avoid repeated sorted when
\code{small_rects} is used as the method of knot selection. This package is
not maintained on CRAN at the time this message is being written; therefore,
we make use of it as a suggested package, following the guidance in
\url{https://cran.r-project.org/doc/manuals/R-exts.html#Suggested-packages}.
I.e., all \code{datastructures} function calls are proceeded with a
namespace and the package is not listed in \code{Imports} or \code{Depends}.
}
\references{
Simon Dirmeier, (2018). datastructures: An R package for organisation and
storage of data . Journal of Open Source Software, 3(28), 910,
\url{https://doi.org/10.21105/joss.00910}

Winston Chang (2020). R6: Encapsulated Classes with Reference Semantics. R
package version 2.5.0. \url{https://CRAN.R-project.org/package=R6}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stepdown-new}{\code{Stepdown$new()}}
\item \href{#method-Stepdown-get_cum_probs}{\code{Stepdown$get_cum_probs()}}
\item \href{#method-Stepdown-get_norm_const}{\code{Stepdown$get_norm_const()}}
\item \href{#method-Stepdown-get_log_x_vals}{\code{Stepdown$get_log_x_vals()}}
\item \href{#method-Stepdown-get_log_h_vals}{\code{Stepdown$get_log_h_vals()}}
\item \href{#method-Stepdown-get_rects}{\code{Stepdown$get_rects()}}
\item \href{#method-Stepdown-get_knot_order}{\code{Stepdown$get_knot_order()}}
\item \href{#method-Stepdown-get_log_p}{\code{Stepdown$get_log_p()}}
\item \href{#method-Stepdown-get_priority_weight}{\code{Stepdown$get_priority_weight()}}
\item \href{#method-Stepdown-add}{\code{Stepdown$add()}}
\item \href{#method-Stepdown-q}{\code{Stepdown$q()}}
\item \href{#method-Stepdown-r}{\code{Stepdown$r()}}
\item \href{#method-Stepdown-d}{\code{Stepdown$d()}}
\item \href{#method-Stepdown-p}{\code{Stepdown$p()}}
\item \href{#method-Stepdown-clone}{\code{Stepdown$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-new}{}}}
\subsection{Method \code{new()}}{
Construct the step function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$new(
  w,
  g,
  tol,
  N,
  method,
  priority_weight = 1/2,
  midpoint_type = "geometric"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{An object representing a weight function (see details).}

\item{\code{g}}{An object representing a base distribution (see details).}

\item{\code{tol}}{A small positive number used in search for \eqn{u_L} and \eqn{u_H}.}

\item{\code{N}}{Number of knot points will be N+1.}

\item{\code{method}}{Either \code{equal_steps} or \code{small_rects} (see details).}

\item{\code{priority_weight}}{Experimental: An weight between 0 and 1. When closer
to 1, more priority is given in \code{small_rects} to rectangle height. When
closer to 0, more priority is given to rectangle width.}

\item{\code{midpoint_type}}{Type of midpoint function to use. Currently only
\code{geometric} or \code{arithmetic} are supported. \code{geometric} is
preferred when the support of \eqn{p(u)} is focused very close to zero.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_cum_probs"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_cum_probs}{}}}
\subsection{Method \code{get_cum_probs()}}{
Get cumulative probabilities after normalizing
the step function to a discrete distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_cum_probs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_norm_const"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_norm_const}{}}}
\subsection{Method \code{get_norm_const()}}{
The constant used to normalize the step function
to a discrete distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_norm_const()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_log_x_vals"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_log_x_vals}{}}}
\subsection{Method \code{get_log_x_vals()}}{
The knot values on which the approximation is based,
returned on the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_x_vals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_log_h_vals"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_log_h_vals}{}}}
\subsection{Method \code{get_log_h_vals()}}{
Values of the approximation evaluated at the points
returned by \code{get_log_x_vals}, given at the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_h_vals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_rects"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_rects}{}}}
\subsection{Method \code{get_rects()}}{
Rectangles used to bound the approximation: x and h
coordinates along with the area. Values are returned on the log-scale or
the original scale according to the \code{log} argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_rects(log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log}}{If \code{TRUE}, return value on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_knot_order"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_knot_order}{}}}
\subsection{Method \code{get_knot_order()}}{
Returns the order in which knot points were added to the
step function. Indices correspond to the points returned by
\code{get_log_x_vals}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_knot_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_log_p"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_log_p}{}}}
\subsection{Method \code{get_log_p()}}{
Evaluate the function \eqn{\rm{Pr}(A_u)} on a given
point u (to be provided on the log-scale). The result is returned
on the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_p(log_u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_u}}{A value of u given at the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-get_priority_weight"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-get_priority_weight}{}}}
\subsection{Method \code{get_priority_weight()}}{
Returns the value of \code{priority_weight}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_priority_weight()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-add"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-add}{}}}
\subsection{Method \code{add()}}{
Update step function by adding a knot at the point u
(given on the log-scale).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$add(log_u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_u}}{A value of u given at the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-q"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-q}{}}}
\subsection{Method \code{q()}}{
Quantiles of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$q(p, log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{A probability to evaluate.}

\item{\code{log}}{If \code{TRUE}, return value on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-r"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-r}{}}}
\subsection{Method \code{r()}}{
Draw from the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$r(n, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of draws to generate.}

\item{\code{log}}{If \code{TRUE}, return value on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-d"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-d}{}}}
\subsection{Method \code{d()}}{
Density of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$d(log_x, log = FALSE, normalize = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_x}}{A vector of points to evaluate, provided on the log-scale.}

\item{\code{log}}{If \code{TRUE}, return value on the log-scale.}

\item{\code{normalize}}{If \code{TRUE}, normalize the result to be a density value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-p"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-p}{}}}
\subsection{Method \code{p()}}{
CDF of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$p(log_x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_x}}{A vector of points to evaluate, provided on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stepdown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stepdown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
