% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Util}
\alias{Util}
\alias{printf}
\alias{logger}
\title{Utilities}
\usage{
printf(msg, ...)

logger(msg, ...)
}
\arguments{
\item{msg}{A string}
}
\description{
Utilities
}
\details{
\code{printf} mimics \code{printf} in C. \code{logger} prepends the message
with a time stamp, which is useful in tracking long programs.
}
