% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Utilities}
\alias{Utilities}
\alias{printf}
\alias{logger}
\title{Utilities}
\usage{
printf(msg, ...)

logger(msg, ...)
}
\arguments{
\item{msg}{A string which can have format placeholders as in \code{sprintf}.}

\item{...}{Arguments to \code{sprintf} to use in the format.}
}
\description{
Utilities
}
\details{
\code{printf} mimics \code{printf} in C. \code{logger} prepends the message
with a time stamp, which is useful in tracking long programs.
}
