% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Specific Direct Samplers}
\alias{Specific Direct Samplers}
\alias{direct_sampler_lognormal_normal}
\alias{direct_sampler_lognormal_dscnorm}
\alias{direct_sampler_lognormal_laplace}
\alias{direct_sampler_lognormal_dgeom}
\alias{direct_sampler_normal_laplace}
\alias{direct_sampler_gamma_laplace}
\title{Specific Direct Samplers}
\usage{
direct_sampler_lognormal_normal(
  n,
  z,
  mu,
  sigma2,
  tau,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)

direct_sampler_lognormal_dscnorm(
  n,
  z,
  mu,
  sigma2,
  tau,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)

direct_sampler_lognormal_laplace(
  n,
  z,
  mu,
  sigma2,
  lambda,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)

direct_sampler_lognormal_dgeom(
  n,
  z,
  mu,
  sigma2,
  rho,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)

direct_sampler_normal_laplace(
  n,
  z,
  mu,
  sigma2,
  lambda,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)

direct_sampler_gamma_laplace(
  n,
  z,
  alpha,
  beta,
  lambda,
  tol,
  N,
  fill_method,
  max_rejections = 0L
)
}
\arguments{
\item{n}{Number of draws to generate.}

\item{z}{A vector of points to evaluate.}

\item{mu}{Mean parameter for weight function.}

\item{sigma2}{standard deviation parameter for weight function.}

\item{tau}{Standard deviation parameter for base distribution.}

\item{tol}{Tolerance for step function approximation in customized sampler.}

\item{N}{Number of knots to use in approximation for \eqn{p(u)}.}

\item{fill_method}{Knot selection method for customized direct sampler.
Can be either \code{equal_steps} or \code{small_rects}.}

\item{max_rejections}{Maximum number of allowed rejections in accept-reject.
If zero (the default), use step function to approximate \eqn{p(u)}, rather
than as an envelope; this will avoid any rejections but yield an approximate
sample.}

\item{lambda}{Scale parameter for base distribution.}

\item{rho}{Probability parameter for base distribution.}
}
\description{
Functions for some specific direct samplers.
}
\details{
Implementations for the following direct samplers are provided:
\itemize{
\item \code{direct_sampler_lognormal_normal}: Lognormal weight function
\eqn{w(x \mid \mu, \sigma^2)} and Normal base distribution
\eqn{g(x \mid 0, \tau^2)}.

\item \code{direct_sampler_lognormal_dscnorm}: Lognormal weight function
\eqn{w(x \mid \mu, \sigma^2)} and Discrete Normal base distribution
\eqn{g(x \mid 0, \tau^2)}.

\item \code{direct_sampler_lognormal_laplace}: Lognormal weight function
\eqn{w(x \mid \mu, \sigma^2)} and Laplace base distribution
\eqn{g(x \mid 0, \lambda)}.

\item \code{direct_sampler_lognormal_dgeom}: Lognormal weight function
\eqn{w(x \mid \mu, \sigma^2)} and Double Geometric base distribution
\eqn{g(x \mid \rho)}.

\item \code{direct_sampler_normal_laplace}: Normal weight function
\eqn{w(x \mid \mu, \sigma^2)} and Laplace base distribution
\eqn{g(x \mid 0, \lambda)}.
}
}
