% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{Stepdown}
\alias{Stepdown}
\title{Stepdown class}
\description{
Stepdown class

Stepdown class
}
\details{
This R6 class represents a step function approximation to the non-increasing
function \eqn{\rm{Pr}(A_u)} for \eqn{u \in [0,1]}.

The object \code{w} represents a unimodal weight function. It is
expected to contain several members:
\itemize{
\item \code{log_c}: the logarithm of the value c, which is the mode of the
weight function.
\item \code{roots(log_a)}: return the roots of the equation
\eqn{\log w(x) - \rm{log_a} = 0}.
\item \code{eval(x, log = TRUE)}: evaluate the function. Return the value
on the log-scale if \code{log = TRUE}.
}

The object \code{g} represents a base distribution. It is
expected to contain two member functions:
\itemize{
\item \code{pr_interval(x1, x2)}: Return \eqn{\rm{Pr}(x1 < X < x2)} under
distribution g.
\item \code{r_truncated(n, x1, x2)}: Take n draws from distribution g
truncated to the open interval \eqn{(x1, x2)}.
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Stepdown$new()}}
\item \href{#method-get_cum_probs}{\code{Stepdown$get_cum_probs()}}
\item \href{#method-get_norm_const}{\code{Stepdown$get_norm_const()}}
\item \href{#method-get_log_x_vals}{\code{Stepdown$get_log_x_vals()}}
\item \href{#method-get_log_h_vals}{\code{Stepdown$get_log_h_vals()}}
\item \href{#method-get_log_p}{\code{Stepdown$get_log_p()}}
\item \href{#method-add}{\code{Stepdown$add()}}
\item \href{#method-q}{\code{Stepdown$q()}}
\item \href{#method-r}{\code{Stepdown$r()}}
\item \href{#method-d}{\code{Stepdown$d()}}
\item \href{#method-p}{\code{Stepdown$p()}}
\item \href{#method-clone}{\code{Stepdown$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct the step function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$new(w, g, tol, N, method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{An object representing a weight function (see details).}

\item{\code{g}}{An object representing a base distribution (see details).}

\item{\code{tol}}{A small positive number used in search for \eqn{u_L} and \eqn{u_H}.}

\item{\code{N}}{Number of knot points will be N+1.}

\item{\code{method}}{Either \code{equal_steps} or \code{small_rects} (see details).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cum_probs"></a>}}
\if{latex}{\out{\hypertarget{method-get_cum_probs}{}}}
\subsection{Method \code{get_cum_probs()}}{
Get cumulative probabilities after normalizing
the step function to a discrete distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_cum_probs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_norm_const"></a>}}
\if{latex}{\out{\hypertarget{method-get_norm_const}{}}}
\subsection{Method \code{get_norm_const()}}{
The constant used to normalize the step function
to a discrete distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_norm_const()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_x_vals"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_x_vals}{}}}
\subsection{Method \code{get_log_x_vals()}}{
The knot values on which the approximation is based,
returned on the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_x_vals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_h_vals"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_h_vals}{}}}
\subsection{Method \code{get_log_h_vals()}}{
Values of the approximation evaluated at the points
returned by \code{get_log_x_vals}, given at the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_h_vals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_p"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_p}{}}}
\subsection{Method \code{get_log_p()}}{
Evaluate the function \eqn{\rm{Pr}(A_u)} on a given
point u (to be provided on the log-scale). The result is returned
on the log-scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$get_log_p(log_u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_u}}{A value of u given at the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Update step function by adding a knot at the point u
(given on the log-scale).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$add(log_u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_u}}{A value of u given at the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-q"></a>}}
\if{latex}{\out{\hypertarget{method-q}{}}}
\subsection{Method \code{q()}}{
Quantiles of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$q(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{A probability to evaluate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r"></a>}}
\if{latex}{\out{\hypertarget{method-r}{}}}
\subsection{Method \code{r()}}{
Draw from the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$r(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of draws to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d"></a>}}
\if{latex}{\out{\hypertarget{method-d}{}}}
\subsection{Method \code{d()}}{
Density of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$d(x, log = FALSE, normalize = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points to evaluate.}

\item{\code{log}}{If \code{TRUE}, return value on the log-scale.}

\item{\code{normalize}}{If \code{TRUE}, normalize the result to be a density value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p"></a>}}
\if{latex}{\out{\hypertarget{method-p}{}}}
\subsection{Method \code{p()}}{
CDF of the distribution based on the step function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$p(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points to evaluate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stepdown$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
